refactorign continues
